requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
apturl==0.5.2
astor==0.8.1
astunparse==1.6.3
attrs==23.2.0
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.4
command-not-found==0.3
configparser==6.0.0
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.8
docker-pycreds==0.4.0
duplicity==0.8.21
exceptiongroup==1.2.0
extract-emails==5.3.3
fasteners==0.14.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.50.0
fsspec==2023.4.0
future==0.18.2
gast==0.5.4
gdown==3.12.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.11.0
httplib2==0.20.2
idna==3.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
keras==3.2.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.6
lockfile==0.12.2
loguru==0.5.3
louis==3.20.0
-e git+https://github.com/DJAlexJ/LRCN-for-Video-Regression.git@f25feccab903a608a13c96105037a883442a1615#egg=lrcnreg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.7
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
openshot-qt==2.5.1
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pafy==0.5.5
pandas==2.2.0
paramiko==2.9.3
pathtools==0.1.2
pexpect==4.8.0
pillow==10.2.0
proglog==0.1.10
promise==2.3
protobuf==3.20.0
psutil==5.9.8
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.18.1
sentry-sdk==1.40.4
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
ssh-import-id==5.11
streamlit==0.74.1
subprocess32==3.5.4
sympy==1.12
systemd-python==234
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.16.0
Theano==1.0.5
threadpoolctl==3.3.0
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.1
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tornado==6.4
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.0
usb-creator==0.3.7
utils==1.0.2
validators==0.22.0
wadllib==1.3.6
wandb==0.10.14
watchdog==0.10.4
webdriver-manager==4.0.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
youtube-dl==2021.12.17
zipp==1.0.0
